discoverer/requirements.txt
more-itertools==8.4.0
pycoax==0.5.0
pyserial==3.4
sliplib==0.5.0
sortedcontainers==2.1.0
wcwidth==0.1.7
